v3: fixes a bug with InitSubscriptions
